moved GenericModal to shared
