+++ ui/components/Performance/PerformanceResults.tsx
@@ -4,7 +4,7 @@ import Moment from 'react-moment';
 import CustomToolbarSelect from '../CustomToolbarSelect';
 import MesheryChart from '../MesheryChart';
 import GrafanaCustomCharts from '../telemetry/grafana/GrafanaCustomCharts';
-import GenericModal from '../General/Modals/GenericModal';
+import GenericModal from '../shared/Modal/GenericModal';
 import { BarChart as BarChartIcon, Info as InfoIcon, Reply as ReplyIcon } from '../icons';
 import fetchPerformanceResults from '@/graphql/queries/PerformanceResultQuery';
 import NodeDetails from './NodeDetails';
